requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.1
attrs==25.4.0
audioop-lts==0.2.2
beautifulsoup4==4.14.2
boto3==1.40.74
botocore==1.40.74
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
comtypes==1.4.13
decorator==5.2.1
docker==7.1.0
docopt==0.6.2
edge-tts==7.2.3
einops==0.8.1
encodec==0.1.1
executing==2.2.1
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
funcy==2.0
google-auth==2.43.0
google-genai==1.50.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
Js2Py==0.74
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
moondream==0.1.1
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
parso==0.8.5
pillow==10.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==2.12.4
pydantic_core==2.41.5
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.2
pyjsparser==2.7.1
PyMsgBox==2.0.1
pyperclip==1.11.0
pypiwin32==223
PyPrind==2.11.3
PyRect==0.2.0
PyScreeze==1.0.1
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
pytweening==1.2.0
pywin32==311
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
s3transfer==0.14.0
safetensors==0.6.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SpeechRecognition==3.14.3
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
suno-bark @ git+https://github.com/suno-ai/bark.git@f4f32d4cd480dfec1c245d258174bc9bde3c2148
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tokenizers==0.21.4
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wcwidth==0.2.14
websockets==15.0.1
yarl==1.22.0
